# genai-engine/staging.openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Arthur GenAI Engine",
        "version": "2.1.0"
    },
    "paths": {
...
        "/users": {
            "post": {
                "tags": [
                    "User Management"
                ],
                "summary": "Create User",
                "description": "Creates a new user with specific roles. The available roles are TASK-ADMIN and CHAT-USER. The 'temporary' field is for indicating if the user password needs to be reset at the first login.",
                "operationId": "create_user_users_post",
                "security": [
                    {
                        "API Key": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "User Management"
                ],
                "summary": "Search Users",
                "description": "Fetch users.",
                "operationId": "search_users_users_get",
                "security": [
                    {
                        "API Key": []
                    }
                ],
                "parameters": [
                    {
                        "name": "search_string",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Substring to match on. Will search first name, last name, email.",
                            "title": "Search String"
                        },
                        "description": "Substring to match on. Will search first name, last name, email."
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/PaginationSortMethod",
                            "description": "Sort the results (asc/desc)",
                            "default": "desc"
                        },
                        "description": "Sort the results (asc/desc)"
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "description": "Page size. Default is 10. Must be greater than 0 and less than 5000.",
                            "default": 10,
                            "title": "Page Size"
                        },
                        "description": "Page size. Default is 10. Must be greater than 0 and less than 5000."
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "description": "Page number",
                            "default": 0,
                            "title": "Page"
                        },
                        "description": "Page number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserResponse"
                                    },
                                    "title": "Response Search Users Users Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/permissions/check": {
            "get": {
                "tags": [
                    "User Management"
                ],
                "summary": "Check User Permission",
                "description": "Checks if the current user has the requested permission. Returns 200 status code for authorized or 403 if not.",
                "operationId": "check_user_permission_users_permissions_check_get",
                "security": [
                    {
                        "API Key": []
                    }
                ],
                "parameters": [
                    {
                        "name": "action",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/UserPermissionAction",
                            "description": "Action to check permissions of."
                        },
                        "description": "Action to check permissions of."
                    },
                    {
                        "name": "resource",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/UserPermissionResource",
                            "description": "Resource to check permissions of."
                        },
                        "description": "Resource to check permissions of."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{user_id}": {
            "delete": {
                "tags": [
                    "User Management"
                ],
                "summary": "Delete User",
                "description": "Delete a user.",
                "operationId": "delete_user_users__user_id__delete",
                "security": [
                    {
                        "API Key": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "User id, not email.",
                            "title": "User Id"
                        },
                        "description": "User id, not email."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{user_id}/reset_password": {
            "post": {
                "tags": [
                    "User Management"
                ],
                "summary": "Reset User Password",
                "description": "Reset password for user.",
                "operationId": "reset_user_password_users__user_id__reset_password_post",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "User Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PasswordResetRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AuthUserRole": {
                "properties": {
                    "id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "composite": {
                        "type": "boolean",
                        "title": "Composite"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "description",
                    "composite"
                ],
                "title": "AuthUserRole"
            },
...
            "CreateUserRequest": {
                "properties": {
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "password": {
                        "type": "string",
                        "title": "Password"
                    },
                    "temporary": {
                        "type": "boolean",
                        "title": "Temporary",
                        "default": true
                    },
                    "roles": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Roles"
                    },
                    "firstName": {
                        "type": "string",
                        "title": "Firstname"
                    },
                    "lastName": {
                        "type": "string",
                        "title": "Lastname"
                    }
                },
                "type": "object",
                "required": [
                    "email",
                    "password",
                    "roles",
                    "firstName",
                    "lastName"
                ],
                "title": "CreateUserRequest"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "PaginationSortMethod": {
                "type": "string",
                "enum": [
                    "asc",
                    "desc"
                ],
                "title": "PaginationSortMethod"
            },
            "PasswordResetRequest": {
                "properties": {
                    "password": {
                        "type": "string",
                        "title": "Password"
                    }
                },
                "type": "object",
                "required": [
                    "password"
                ],
                "title": "PasswordResetRequest"
            },
...
            "UserPermissionAction": {
                "type": "string",
                "enum": [
                    "create",
                    "read"
                ],
                "title": "UserPermissionAction"
            },
            "UserPermissionResource": {
                "type": "string",
                "enum": [
                    "prompts",
                    "responses",
                    "rules",
                    "tasks"
                ],
                "title": "UserPermissionResource"
            },
            "UserResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "first_name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "First Name"
                    },
                    "last_name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Last Name"
                    },
                    "roles": {
                        "items": {
                            "$ref": "#/components/schemas/AuthUserRole"
                        },
                        "type": "array",
                        "title": "Roles"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "email",
                    "roles"
                ],
                "title": "UserResponse"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "API Key": {
                "type": "http",
                "description": "Bearer token authentication with an API key",
                "scheme": "bearer"
            }
        }
    }
}